LWC support for FQuat and FTransform Merge Dev-LargeWorldCoordinates @ 17289670
#preflight 6125fb6925057900014f2cd6

[CL 17300515 by Andrew Davidson in ue5-main branch]

+++ Engine/Source/Developer/Datasmith/DatasmithFacade/Private/DatasmithFacadeActorLight.cpp
@@ -179,10 +179,10 @@ void FDatasmithFacadeActorLight::GetIesRotation(
 ) const
 {
 	FQuat IesQuat(GetDatasmithLightActorElement()->GetIesRotation());
-	OutX = IesQuat.X;
-	OutY = IesQuat.Y;
-	OutZ = IesQuat.Z;
-	OutW = IesQuat.W;
+	OutX = (float)IesQuat.X;
+	OutY = (float)IesQuat.Y;
+	OutZ = (float)IesQuat.Z;
+	OutW = (float)IesQuat.W;
 }
 
 void FDatasmithFacadeActorLight::GetIesRotation(
